refactor(watch): plugin-based capability system (#708)
Replace monolithic watch.ts (1562 lines) with a plugin architecture
where each opt-in feature is a self-contained WatchCapability.

Architecture:
- watch/types.ts: WatchCapability interface, WatchContext, WatchPhase
- watch/registry.ts: CapabilityRegistry (register, getByPhase, all)
- watch/config.ts: loadWatchConfig merges .squad/config.json + CLI flags
- watch/capabilities/: 9 capability plugins extracted from watch.ts

Capabilities extracted:
- self-pull: git fetch/pull at round start (pre-scan phase)
- execute: spawn Copilot sessions for eligible issues (post-execute)
- board: project board lifecycle + reconciliation (post-execute)
- monitor-teams: Teams scanning via WorkIQ (housekeeping)
- monitor-email: email scanning + GitHub alerts (housekeeping)
- two-pass: lightweight scan then hydrate actionable (post-triage)
- wave-dispatch: parallel sub-task execution (post-execute)
- retro: retrospective enforcement (housekeeping)
- decision-hygiene: merge inbox cleanup (housekeeping)

Config system:
- .squad/config.json 'watch' section for persistent config
- Priority: CLI flag > config.json > default (off)
- --no-{capability} disables even if config enables it
- Auto-generates --help from registered capability descriptions

Backward compatibility:
- Legacy WatchOptions type still accepted by runWatch
- 'squad watch' with no flags = triage only (unchanged)
- 'squad watch --execute' still works as CLI flag
- All existing exports preserved via re-exports

The main loop is now a thin 4-phase orchestrator:
1. pre-scan → 2. core triage → 3. post-execute → 4. housekeeping

Each capability has try/catch in execute() — failures are warnings,
never crash the round. Capabilities are independently testable.
